Create community page updates, Cosmos/DAO support. (#1551)
* Convert Create Community Modal to Page

* Add margin for Save btn, rename top-level classes, inline toggle styles

* Remove unused imports

* Return form inside of a sublayout

* Separate create community tabs into pages.

* Remove doubled substrate button.

* Add ethereum as always available for offchain community.

* Add generic chain input helper.

* Add styling to get correct content width

* Add useQuickSwitcher to Sublayout & Sidebar

* Autogenerate tabs.

* Add icon uploader.

* Clean up button behavior.

* Move substrate test button.

* Cosmos tab.

* Populate height field from block query.

* Support ETH DAO testing.

* Fix up ERC20 creation and add dropdown.

* Fix slugs.

* Remove logs and unnecessary ternary.

* Add case sensitive and remove bad height display.
