+++ client/scripts/models/ChainInfo.ts
@@ -88,7 +88,7 @@ class ChainInfo {
     this.collapsedOnHomepage = collapsedOnHomepage;
     this.defaultSummaryView = defaultSummaryView;
     this.featuredTopics = featuredTopics || [];
-    this.topics = topics.map((t) => new OffchainTopic(t)) || [];
+    this.topics = topics ? topics.map((t) => new OffchainTopic(t)) : [];
     this.adminsAndMods = adminsAndMods || [];
     this.type = type;
     this.ss58Prefix = ss58_prefix;
