+++ client/scripts/views/components/avatar_upload.ts
@@ -98,8 +98,8 @@ const AvatarUpload: m.Component<IAttrs, IState> = {
     const logoURL =
       vnode.state.dropzone?.option?.url ||
       (app.activeCommunityId()
-        ? app.community.meta.iconUrl
-        : app.chain.meta.chain.iconUrl);
+        ? app.community?.meta.iconUrl
+        : app.chain?.meta.chain.iconUrl);
     return m('form.AvatarUpload', [
       m(
         '.dropzone-attach',
