+++ client/scripts/views/components/sidebar/index.ts
@@ -17,7 +17,7 @@ import SubscriptionButton from 'views/components/subscription_button';
 import ChainStatusIndicator from 'views/components/chain_status_indicator';
 import { ChainIcon, CommunityIcon } from 'views/components/chain_icon';
 import CommunitySelector from 'views/components/sidebar/community_selector';
-import CreateCommunityModal from 'views/modals/create_community_modal';
+
 
 import { DiscordIcon, TelegramIcon, ElementIcon, GithubIcon, WebsiteIcon } from '../component_kit/icons';
 
@@ -81,7 +81,8 @@ const SidebarQuickSwitcher: m.Component<{}> = {
           rounded: true,
           label: m(Icon, { name: Icons.PLUS }),
           onclick: (e) => {
-            app.modals.create({ modal: CreateCommunityModal });
+            e.preventDefault();
+            m.route.set('/createCommunity');
           },
         }),
       ]),
@@ -604,11 +605,13 @@ export const ExternalLinksModule: m.Component<{}, {}> = {
   }
 };
 
-const Sidebar: m.Component<{ hideQuickSwitcher? }, {}> = {
+const Sidebar: m.Component<{ hideQuickSwitcher?, useQuickSwitcher?: boolean }, {}> = {
   view: (vnode) => {
+    const { useQuickSwitcher } = vnode.attrs;
+
     return [
       !app.isCustomDomain() && m(SidebarQuickSwitcher),
-      m('.Sidebar', [
+      !useQuickSwitcher && m('.Sidebar', [
         (app.chain || app.community) && m(OffchainNavigationModule),
         (app.chain || app.community) && m(OnchainNavigationModule),
         (app.chain || app.community) && m(ExternalLinksModule),
