--- client/styles/modals/create_community_modal.scss
@@ -1,45 +1,0 @@
-@import 'client/styles/shared';
-
-.CreateCommunityModal {
-    padding: 6px 28px 10px;
-
-    h3 {
-        text-align: center;
-        margin: 24px 0 20px;
-    }
-    h4 {
-        margin: 4px 0;
-    }
-    p {
-        text-align: center;
-    }
-    input[type="text"] {
-        @include formfield-text;
-        margin: 10px 0;
-    }
-    button, a.btn {
-        display: block;
-        width: 100%;
-        text-align: center;
-    }
-    form {
-        margin: 10px auto 20px;
-    }
-    .auth-features {
-        margin: 10px 0 26px;
-        width: 440px;
-    }
-    .create-community-message {
-        text-align: center;
-        margin: 20px 0;
-        &.error {
-            @include error-text();
-        }
-        &.success {
-            @include success-text();
-        }
-    }
-}
-.CompactModalExitButton {
-    color: $text-color-medium;
-}
